pdf2jpeg/requirements.txt
attrs==19.3.0
awscli==1.18.32
boto3==1.12.33
botocore==1.15.33
chalice==1.13.0
click==7.1.1
colorama==0.4.3
docutils==0.15.2
enum-compat==0.0.3
jmespath==0.9.5
pdf2image==1.12.1
Pillow==7.0.0
pyasn1==0.4.8
python-dateutil==2.8.1
PyYAML==5.3.1
rsa==3.4.2
s3transfer==0.3.3
six==1.14.0
urllib3==1.25.8
